Update schema generation class
